pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trishift"
version = "0.1.0"
description = "TriShift package"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch",
  "numpy",
  "pandas",
  "anndata",
  "scanpy",
  "scipy",
  "pyyaml",
  "tqdm",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
